feat: extract `DecimalU256` to `x402-types` and clean up duplicate code
- Moved `DecimalU256` implementation to `x402-types` for shared usage.
- Updated imports to use the centralized implementation.
- Removed redundant `DecimalU256` code from `x402-chain-eip155`.
