+++ crates/x402-types/src/util/mod.rs
@@ -7,7 +7,9 @@
 //! - [`money_amount`] - Human-readable currency amount parsing
 
 pub mod b64;
+pub mod decimal_u256;
 pub mod lit_str;
 pub mod money_amount;
 
 pub use b64::*;
+pub use decimal_u256::*;
